Allow skipping Vercel builds by using specific tags in commit messages
Examples of tags that will skip the builds:

- `[no deploy]`
- `[skip Vercel]`
- `[skip examples]`

Basically any combination of: `[{no,skip} {deploy,examples,vercel}]` is
recognized (case-insensitive).

Once such a commit lands on this PR or branch, there is no way to "undo"
that, unless you rewrite the Git history by removing the commit or
changing the commit message.