Reduce runtime type check to only check for strings
This is done to remain forward-compatible.

Over time, we may want to expand the supported scopes, and if the
backend ever starts emitting a new scope in issued tokens, we don't want
old clients to choke on those just because that token is unknown to
them.

We'll still expose the `isScope` check, to enable easy filtering.