Change `.toJson()` to `.toImmutable()`
This is a subtle one, especially because it leads to a bit of
duplication at first sight. The thing is that if we map our LSON values
like so:

  LiveMap<string, V>          ->   { [string]: V }
  LiveObject<{ x: number }>   ->   { x: number }

Then there's no way to find a reverse mapping for these at the type
level.  This would be needed for implementing a `useMutable()` API.

For example:

   useMutable(root => root.x.y.z)

If `z` here is a JavaScript object, it's unclear whether that is the
immutable representatino of a `LiveMap` or a `LiveObject`, and the best
shot we have would be to return a union type... :(

However, if we map the LSON types as follows:

  LiveMap<string, V>          ->   ReadonlyMap<string, V>
  LiveObject<{ x: number }>   ->   { readonly x: number }

Then we would be able to reverse-map them!

  ReadonlyMap<string, V>      ->   LiveMap<string, V>
  { readonly x: number }      ->   LiveObject<{ x: number }>

Besides, correlating `LiveMap` more directly with ES6 `Map`s makes a lot
of sense anyway!