Fix auto-detection of external packages in rollup configs
This fixes an issue where, if you had `some-pkg` declared in your
dependencies, but then imported from `"some-pkg/sub/module"`, then
Rollup would not consider that an external dependency, because it was
not an exact match.

The fix is to not use a static list, but to make `external` a function
that can do a prefix-match.

This fix also allows us to get rid of the custom
`@liveblocks/client/internal` hack we had in there to support a similar
case.