Fix suboptimal behavior of useOther
This makes `useOther` return the last known value instead of throwing an
error when used with a non-existing connection ID. This helps to prevent
a problem where the children are rerendered before the parent got the
chance to unmount them.

Effectively, this means that "cursors stop getting rerendered for users
that left the room", assuming they will get unmounted shortly after by
the controlling parent component.