Add new `backgroundKeepAliveTimeout` client option
It's optional. Without setting it, the client will forever try to keep
the connection to the Liveblocks server alive, just like it does today.

When specifying a number of milliseconds, clients are allowed to go into
a dorment "zombie state", where they are disconnected from the server as
long as they remain in the background (so not permanently disconnected).
When such tabs become visible again, they will instantly automatically
reconnect and sync up again.