Remove all access of process.env.* (except NODE_ENV)
The reason is that Vite does not seem to rewrite any of these in the
final bundle, leaving those `process.env.WHATEVER` patterns in the final
bundle.

Other bundlers do rewrite these expressions to `undefined`, which is
what the expected result was here.

It's still fine to leave `process.env.NODE_ENV`, as that specific
pattern _is_ recognized by Vite (and all other bundlers), so it's still
fine to use that.