Remove `use client` banner
This didn't end up in the final bundle as a directive anyway, but as
a string literal instead, meaning it was not doing anything before
already.

It likely wasn't useful anyway, since users have to specify "use client"
in their own component files anyway if they want to use our hooks?