Simplify createBatchStore implementation
Only expose read-only Observable API on the outside, no need to expose
the `.notify()` API.

Also, removed `undefined` as a possible internal state.

Do not send state as payload to notify events, make it just a void
notifier. The reason is that it's confusing to send an event without
also sending which key got updated with that state.