Move all existing fetch helpers into HttpClient
This is still very ugly as they are all pretty similar and offer no good
abstraction just yet, but it's the first step to seeing the to-be-DRY'ed
up parts, and to extract the commonalities.

Also, the subtle difference between http client 1 and 2 is becoming
more obvious now.

More refactoring needed.