Remove various `.subscribeXxx()` methods that are all aliases to .subscribe()
The intention was to split these out by use case, but it's going to take
a bit more refactoring to get there. Until we can do that, it's less
confusing if they all just use a global "subscribe", which is closer to
how it's actually working at this point still.