Recognize a third (hidden) sync state: has-local-changes
This state is a nuance between "synchronizing" and "synchronized" and
will decouple the following two things:

1. The `useSyncStatus()` hook will be either "synchronizing" or
   "synchronized" (to show a badge)
2. Browser tabs will be prevented from closing if there are unsaved
   changes.

The use case that's in between these two things is having draft text (or
attachment uploads) in a new comment box. In that state, we _do_ want to
prevent the tab from closing, but not display a "synchronizing" spinner.
We'll only want to show that spinner once the user submits the
attachment.