For now, allow description only on success results
Adding it in addition to the error texts or cancellation types might be
a bit confusing actually at the moment. We can revert this decision
later and open up the possibility for this again, but not the other way
around, so this is the least risky option to launch as V1.