Implement alternative stateful partial JSON parser
This stateful parser is more efficinet at runtime because it keeps the
parsed state so far in memory and does not have to perform the same
parsing of every prefix of the string again and again as new content
arrives at the end.

Also, this implementation covers more edge cases.